pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "My third project"
authors = ["Antesser"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
